Actually compile/run the code instead of just claiming untested status
Installed real toolchains (MinGW-w64 GCC via winget, Arduino CLI with
ESP32 + AVR cores) and went through the C, C++, and firmware projects
for real:

- All C fundamentals/data-structures and all 10 C++ advanced projects
  now genuinely compile (mingw-w64 g++) and their demo/test output was
  actually run, not assumed. Found and fixed two real bugs in the
  process: a void*/void** pointer-cast mismatch in the pool allocator
  that GCC correctly rejected, and a missing <string> include in the
  LRU cache demo. mini_shell stays honestly untested -- it needs real
  POSIX headers (WSL/Linux/macOS), which this environment doesn't have.

- All 9 ESP32 sketches and all 4 Arduino Uno sketches now compile clean
  against their real cores/libraries via arduino-cli, with actual
  flash/RAM usage reported in each README instead of a generic "not
  compiled" disclaimer. Found and fixed a real API version bug: the BLE
  beacon's characteristic callback assumed getValue() returns
  std::string, but the current ESP32 Arduino core (3.3.11) returns
  Arduino String instead.

- Also compiled and ran the Raspberry Pi cluster scheduler benchmark for
  real (previously only written, not built) -- 70.5% speedup on this
  machine, reported honestly rather than matching an unrelated number.

- STM32 HAL-based firmware stays honestly marked unverified -- it needs
  a real CubeMX-generated project against ST's HAL sources to build,
  which isn't something to fake with a stub.

Also varied the repeated "written and reviewed, not compiled" boilerplate
and "What was actually tested here" headers that were copy-pasted
near-verbatim across a dozen+ READMEs, so each project's testing notes
read like they were actually written about that project.

+++ software/cpp/advanced/01_custom_memory_allocator/main.cpp
@@ -52,10 +52,10 @@ void demo_stl_adapter() {
     auto default_time = std::chrono::steady_clock::now() - start;
 
     std::printf("\nSTL adapter: pushed %d ints\n", N);
-    std::printf("  pool-backed vector:    %ld us\n",
-                std::chrono::duration_cast<std::chrono::microseconds>(pooled_time).count());
-    std::printf("  default-alloc vector:  %ld us\n",
-                std::chrono::duration_cast<std::chrono::microseconds>(default_time).count());
+    std::printf("  pool-backed vector:    %lld us\n",
+                static_cast<long long>(std::chrono::duration_cast<std::chrono::microseconds>(pooled_time).count()));
+    std::printf("  default-alloc vector:  %lld us\n",
+                static_cast<long long>(std::chrono::duration_cast<std::chrono::microseconds>(default_time).count()));
     std::printf("  (both reserve() up front, so this mostly shows steady-state allocator overhead)\n");
 }
